requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2
django-extensions==2.1.6
django-heroku==0.3.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-yasg==1.15.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
Pillow==6.0.0
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
PyJWT==1.7.1
pytest==4.4.0
pytest-django==3.4.8
python-dotenv==0.10.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
whitenoise==4.1.2
